requirements.txt
ansiwrap==0.8.4
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
configparser==5.2.0
cycler==0.11.0
DateTime==4.5
debuglater==1.4.3
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dotenvy==0.2.0
eikon==1.1.15
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.9.1
fastjsonschema==2.16.1
fonttools==4.34.4
future==0.18.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.3.0
hyperframe==5.2.0
idna==2.10
importlib-resources==5.9.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupytext==1.14.1
kiwisolver==1.4.4
lxml==4.9.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mdurl==0.1.2
mistune==0.8.4
monotonic==1.6
nbclient==0.6.6
nbconvert==6.5.1
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
patsy==0.5.2
pd==0.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
ploomber==0.21
ploomber-core==0.0.6
ploomber-engine==0.0.10
ploomber-scaffold==0.3.1
pluggy==1.0.0
posthog==2.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
pyflakes==2.5.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyvis==0.2.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.8.17
requests==2.28.1
rfc3986==1.5.0
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data==0.3.0
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.0.1
terminado==0.15.0
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
widgetsnbextension==3.6.1
yellowbrick==1.5
zipp==3.8.1
zope.interface==5.4.0
